Split visualizer into multiple coherent files
