+++ CMakeLists.txt
@@ -169,7 +169,8 @@ add_library(ql
     "${CMAKE_CURRENT_SOURCE_DIR}/src/clifford.cc"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/passmanager.cc"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/passes.cc"
-    "${CMAKE_CURRENT_SOURCE_DIR}/src/visualizer.cc"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/visualizer_common.cc"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/visualizer_circuit.cc"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/report.cc"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/exception.cc"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/arch/cc/eqasm_backend_cc.cc"
