+++ src/visualizer_circuit.h
@@ -1,39 +1,25 @@
 /**
- * @file   visualizer_internal.h
- * @date   09/2020
+ * @file   visualizer_circuit.h
+ * @date   11/2020
  * @author Tim van der Meer
- * @brief  declaration of the visualizer's internals
+ * @brief  definition of the visualizer
  */
 
 #pragma once
 
 #ifdef WITH_VISUALIZER
-
+ 
 #include "visualizer.h"
+#include "visualizer_common.h"
 #include "CImg.h"
-
-// These undefs are necessary to avoid name collisions between CImg and Lemon.
-#undef cimg_use_opencv
-#undef True
-
-#undef False
-#undef IN
-#undef OUT
 
 namespace ql {
 
-enum BitType {CLASSICAL, QUANTUM};
-
-struct Position4 {
-    long x0 = 0;
-    long y0 = 0;
-    long x1 = 0;
-    long y1 = 0;
-};
-
-struct Position2 {
-    long x = 0;
-    long y = 0;
+struct Cycle {
+    int index = 0;
+    bool empty = false;
+    bool cut = false;
+    std::vector<std::vector<std::reference_wrapper<GateProperties>>> gates;
 };
 
 struct Cell {
@@ -41,49 +27,6 @@ struct Cell {
     const int row;
     const int chunkOffset;
     const BitType bitType;
-};
-
-struct EndPoints {
-    const int start;
-    const int end;
-};
-
-struct Dimensions {
-    const int width;
-    const int height;
-};
-
-struct GateOperand {
-    BitType bitType = QUANTUM;
-    int index = 0;
-
-    friend bool operator<(const GateOperand &lhs, const GateOperand &rhs) {
-        if (lhs.bitType == QUANTUM && rhs.bitType == CLASSICAL) return true;
-        if (lhs.bitType == CLASSICAL && rhs.bitType == QUANTUM) return false;
-        return lhs.index < rhs.index;
-    }
-
-    friend bool operator>(const GateOperand &lhs, const GateOperand &rhs) {return operator<(rhs, lhs);}
-    friend bool operator<=(const GateOperand &lhs, const GateOperand &rhs) {return !operator>(lhs, rhs);}
-    friend bool operator>=(const GateOperand &lhs, const GateOperand &rhs) {return !operator<(lhs, rhs);}
-};
-
-struct GateProperties {
-    std::string name;
-    std::vector<int> operands;
-    std::vector<int> creg_operands;
-    int duration = 0;
-    int cycle = 0;
-    gate_type_t type = __custom_gate__;
-    std::vector<int> codewords;
-    std::string visual_type;
-};
-
-struct Cycle {
-    int index = 0;
-    bool empty = false;
-    bool cut = false;
-    std::vector<std::vector<std::reference_wrapper<GateProperties>>> gates;
 };
 
 enum LineSegmentType {FLAT, PULSE, CUT};
@@ -198,24 +141,13 @@ class Structure {
     void printProperties() const;
 };
 
-Layout parseConfiguration(const std::string &configPath);
+void visualizeCircuit(std::vector<GateProperties> gates, const Layout layout, const int cycleDuration, const std::string &waveformMappingPath);
+
 PulseVisualization parseWaveformMapping(const std::string &waveformMappingPath);
-void validateLayout(Layout &layout);
-
-std::vector<GateProperties> parseGates(const ql::quantum_program* program);
-
-int calculateAmountOfGateOperands(const GateProperties gate);
-std::vector<GateOperand> getGateOperands(const GateProperties gate);
-std::pair<GateOperand, GateOperand> calculateEdgeOperands(const std::vector<GateOperand> operands, const int amountOfQubits);
-
-void fixMeasurementOperands(std::vector<GateProperties> &gates);
-bool isMeasurement(const GateProperties gate);
 
 std::vector<QubitLines> generateQubitLines(const std::vector<GateProperties> gates, const PulseVisualization pulseVisualization, const CircuitData circuitData);
 double calculateMaxAmplitude(const std::vector<LineSegment> lineSegments);
 void insertFlatLineSegments(std::vector<LineSegment> &existingLineSegments, const int amountOfCycles);
-
-Dimensions calculateTextDimensions(const std::string &text, const int fontHeight, const Layout layout);
 
 void drawCycleLabels(cimg_library::CImg<unsigned char> &image, const Layout layout, const CircuitData circuitData, const Structure structure);
 void drawCycleEdges(cimg_library::CImg<unsigned char> &image, const Layout layout, const CircuitData circuitData, const Structure structure);
@@ -235,9 +167,6 @@ void drawGateNode(cimg_library::CImg<unsigned char> &image, const Layout layout,
 void drawControlNode(cimg_library::CImg<unsigned char> &image, const Layout layout, const Structure structure, const Node node, const Cell cell);
 void drawNotNode(cimg_library::CImg<unsigned char> &image, const Layout layout, const Structure structure, const Node node, const Cell cell);
 void drawCrossNode(cimg_library::CImg<unsigned char> &image, const Layout layout, const Structure structure, const Node node, const Cell cell);
-
-void printGates(const std::vector<GateProperties> gates);
-int safe_int_cast(const size_t argument);
 
 } // namespace ql
