+++ tests/visualizer/visualizer_config_example.json
@@ -75,7 +75,7 @@
     },
     "pulses":
     {
-        "displayGatesAsPulses": false,
+        "displayGatesAsPulses": true,
         "pulseRowHeightMicrowave": 32,
         "pulseRowHeightFlux": 32,
         "pulseRowHeightReadout": 32,
